app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.1",
    "anthropic>=0.30.0",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "pypdf>=6.1.3",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
]
